Allow `ReverseProxy.client` to support other connectors than `HttpConnector`
This is part 1 of a series of commits to address issue #43

The problem is that an `HttpConnector` does not support TLS connections
so attempting to proxy to an https URI causes the proxy to connect to
the right port but attempt to talk HTTP, not HTTPS. Obviously that'll
never work.

With this chance a consumer of this lib can pass an `HttpsConnection`
from hyper-tls or any other type that implements `Connect`.