Introduce tls feature flag and make it a default flag
This is part 2 of a series of commits to address issue #43

When set we'll add a dependency on `hyper-tls` and always use an
`Client<HttpsConnector, _>` in the reverse proxy. This way we
support both HTTP and HTTPS. It's a feature because users of the
library may not want to depend on `hyper-tls` for various reasons.

If so they can simply not set the flag and not use TLS or use
any TLS capable connector they please.

Note that this change fixes the examples.